+++ Patches/GameBinding.cs
@@ -87,6 +87,7 @@ namespace HumankindAssetFramework
         internal static Type AbstractAnimationStateMachine => Cached("Amplitude.Mercury.Presentation.AbstractAnimationStateMachine");
         internal static Type PresentationArtilleryStrikeController => Cached("Amplitude.Mercury.Presentation.PresentationArtilleryStrikeController");
         internal static Type PresentationArtilleryStrike => Cached("Amplitude.Mercury.Presentation.PresentationArtilleryStrike");
+        internal static Type UnitActionRangedFightSequence => Cached("Amplitude.Mercury.Presentation.UnitActionRangedFightSequence");
         internal static Type WorldPosition          => Cached("Amplitude.Mercury.WorldPosition");
         internal static Type WorldPositionExtensions => Cached("WorldPositionExtensions", "Amplitude.Mercury.WorldPositionExtensions");
         internal static Type UnitActionFaceEnemy      => Cached("Amplitude.Mercury.Presentation.UnitActionFaceEnemy");
@@ -194,6 +195,7 @@ namespace HumankindAssetFramework
             new Dep(AbstractAnimationStateMachine, nameof(AbstractAnimationStateMachine), "StepWaitingDelay", "ownerPawn"),
             new Dep(PresentationArtilleryStrikeController, nameof(PresentationArtilleryStrikeController), "ScheduleArtilleryStrikeProjectileLaunch", "ScheduleArtilleryStrikeHit"),
             new Dep(PresentationArtilleryStrike, nameof(PresentationArtilleryStrike), "TriggerArtilleryStrikeVisuals", "TriggerArtilleryStrikeFX", "PrepareArtilleryStrikeFX", "projectileData", "TargetTileIndex", "AttackerArmyGUID"),
+            new Dep(UnitActionRangedFightSequence, nameof(UnitActionRangedFightSequence), "AddPawnRangedFightSequence"),
             new Dep(WorldPosition, nameof(WorldPosition)),
             new Dep(WorldPositionExtensions, nameof(WorldPositionExtensions), "ToVector3"),
             new Dep(UnitActionFaceEnemy, nameof(UnitActionFaceEnemy), "StartUnitAction", "actionScope", "AttackerBattleUnit"),